clean up and add css
