+++ front-end/nycroute-app/src/SubwayLinesInfo.css
@@ -1,10 +1,19 @@
+.SubwayLinesInfo{
+    margin: 10px;
+    margin-top: 30px;
+}
+
+h1{
+    text-align: center;
+}
+
 #regularOrder{
     display: flex;
     flex-direction: column;
- }
+}
  
- #reverseOrder{
-     display: flex;
-     flex-direction: column-reverse;
- }
+#reverseOrder{
+    display: flex;
+    flex-direction: column-reverse;
+}
