+++ front-end/nycroute-app/src/SubwayLinesInfoItem.css
@@ -1,0 +1,9 @@
+li{
+    list-style-type: none;
+}
+
+.list{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
